# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fitrack",
        "description": "Documentación de la API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/target-locations": {
            "get": {
                "tags": [
                    "TargetLocation"
                ],
                "summary": "Obtener todas las ubicaciones objetivo",
                "description": "Devuelve una lista de todas las entidades de ubicación objetivo.",
                "operationId": "getAllTargetLocations",
                "responses": {
                    "200": {
                        "description": "Lista de ubicaciones objetivo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Casa"
                                            },
                                            "position": {
                                                "properties": {
                                                    "latitude": {
                                                        "type": "number",
                                                        "example": 19.4326
                                                    },
                                                    "longitude": {
                                                        "type": "number",
                                                        "example": -99.1332
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "radiusMeters": {
                                                "type": "number",
                                                "example": 50
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/target-locations/{user_id}": {
            "post": {
                "tags": [
                    "TargetLocation"
                ],
                "summary": "Crear nueva ubicación objetivo",
                "description": "Crea una nueva ubicación objetivo para un usuario específico.",
                "operationId": "createTargetLocation",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "position",
                                    "radiusMeters"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Casa"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "19.4326,-99.1332"
                                    },
                                    "radiusMeters": {
                                        "type": "number",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ubicación creada exitosamente"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/target-locations/{user_id}/{id}": {
            "delete": {
                "tags": [
                    "TargetLocation"
                ],
                "summary": "Eliminar ubicación objetivo de un usuario",
                "description": "Elimina una ubicación objetivo específica asociada a un usuario dado.",
                "operationId": "deleteTargetLocation",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ubicación eliminada exitosamente"
                    },
                    "404": {
                        "description": "Ubicación no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token en formato Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}